backend/requirements.txt
# Core Django & DRF
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
django-filter>=24.2
django-guardian>=2.4.0
django-tenants>=3.6.1

# Database & Storage
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
django-storages[google,s3]>=1.14.2
whitenoise>=6.6.0

# Async & Realtime
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.1.0
celery>=5.3.6
redis>=5.0.3
django-redis>=5.4.0

# AI & Automation
openai>=1.14.3

# Utilities
django-environ>=0.11.2
python-dotenv>=1.0.1
django-extensions>=3.2.3
django-import-export>=3.3.6
openpyxl>=3.1.2
tablib>=3.5.0
pillow>=10.2.0

# Documentation
drf-spectacular>=0.27.1

# Production
gunicorn>=21.2.0
uvicorn>=0.29.0
sentry-sdk>=1.43.0

# Testing
pytest>=8.1.1
pytest-django>=4.8.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=24.3.0
